# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Enterprise Deal Commander (EDC) Phase 1 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all commander accounts
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "403":
          description: Forbidden (reader)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Change a user's display name, role, or active state (admin only)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "403":
          description: Forbidden (reader)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: >
            Self-demotion/deactivation, or this is the last active admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Permanently delete a user (admin only)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "403":
          description: Forbidden (reader)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Self-delete, or this is the last active admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/users/{id}/password:
    post:
      operationId: resetUserPassword
      tags: [users]
      summary: Reset a user's password (admin only; also usable on self)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Password too short
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden (reader)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ============ DEALS ============
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message: { type: string }
      required: [message]
    ErrorDetail:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        details:
          type: array
          items: { type: object, additionalProperties: true }
      required: [code, message]
    ErrorResponse:
      type: object
      properties:
        error: { $ref: "#/components/schemas/ErrorDetail" }
      required: [error]
...
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        displayName: { type: string }
        role: { type: string, enum: [admin, reader] }
        isActive: { type: boolean }
        createdAt: { type: string }
        lastDashboardVisitAt: { type: ["string", "null"] }
      required: [id, email, displayName, role, isActive, createdAt, lastDashboardVisitAt]
    UserListResponse:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/User" }
      required: [data]
    UserResponse:
      type: object
      properties:
        data: { $ref: "#/components/schemas/User" }
      required: [data]
    CreateUserInput:
      type: object
      properties:
        email: { type: string, minLength: 1 }
        display_name: { type: string, minLength: 1, maxLength: 255 }
        password: { type: string, minLength: 12 }
        role: { type: string, enum: [admin, reader] }
      required: [email, display_name, password]
    UpdateUserInput:
      type: object
      properties:
        display_name: { type: string, minLength: 1, maxLength: 255 }
        role: { type: string, enum: [admin, reader] }
        is_active: { type: boolean }
      minProperties: 1
    ResetPasswordInput:
      type: object
      properties:
        password: { type: string, minLength: 12 }
      required: [password]
...