# lib/api-zod/src/generated/types/createUserInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Enterprise Deal Commander (EDC) Phase 1 API
 * OpenAPI spec version: 1.0.0
 */
import type { CreateUserInputRole } from './createUserInputRole';

export interface CreateUserInput {
  /** @minLength 1 */
  email: string;
  /**
     * @minLength 1
     * @maxLength 255
     */
  display_name: string;
  /** @minLength 12 */
  password: string;
  role?: CreateUserInputRole;
}