# lib/api-zod/src/generated/types/updateUserInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Enterprise Deal Commander (EDC) Phase 1 API
 * OpenAPI spec version: 1.0.0
 */
import type { UpdateUserInputRole } from './updateUserInputRole';

export interface UpdateUserInput {
  /**
     * @minLength 1
     * @maxLength 255
     */
  display_name?: string;
  role?: UpdateUserInputRole;
  is_active?: boolean;
}