# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Enterprise Deal Commander (EDC) Phase 1 API
 * OpenAPI spec version: 1.0.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: string;
  email: string;
  displayName: string;
  role: UserRole;
  isActive: boolean;
  createdAt: string;
  /** @nullable */
  lastDashboardVisitAt: string | null;
}