# lib/api-zod/src/generated/types/userResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Enterprise Deal Commander (EDC) Phase 1 API
 * OpenAPI spec version: 1.0.0
 */
import type { User } from './user';

export interface UserResponse {
  data: User;
}